Separate module for cavoke developer lib. Move that module to the server's docker
