+++ cavoke-dev-lib/cavoke.cpp
