+++ server/Dockerfile
@@ -8,6 +8,8 @@ EXPOSE $SERVER_PORT
 # Cavoke server
 ENV CAVOKE_ROOT="$IROOT/cavoke/"
 ADD ./ $CAVOKE_ROOT
+# add cavoke-dev-lib dependecy
+ADD ../cavoke-dev-lib $CAVOKE_ROOT
 WORKDIR $CAVOKE_ROOT
 # Install server
 RUN chmod +x ./install.sh && ./install.sh
